x64-exp
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/64bit-exp@1846 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: 96e5d46cae25a5c9392fe926f8df35002f360037

+++ core/consumer/frame_consumer.h
@@ -44,7 +44,7 @@ struct frame_consumer : boost::noncopyable
 	virtual std::wstring print() const = 0;
 	virtual boost::property_tree::wptree info() const = 0;
 	virtual bool has_synchronization_clock() const {return true;}
-	virtual size_t buffer_depth() const = 0;
+	virtual int buffer_depth() const = 0;
 	virtual int index() const = 0;
 
 	static const safe_ptr<frame_consumer>& empty();
