+++ modules/ffmpeg/producer/audio/audio_resampler.cpp
@@ -46,13 +46,13 @@ struct audio_resampler::implementation
 	std::vector<int8_t, tbb::cache_aligned_allocator<int8_t>> copy_buffer_;
 	std::vector<int8_t, tbb::cache_aligned_allocator<int8_t>> buffer2_;
 
-	const size_t			output_channels_;
+	const int			output_channels_;
 	const AVSampleFormat	output_sample_format_;
 
-	const size_t			input_channels_;
+	const int			input_channels_;
 	const AVSampleFormat	input_sample_format_;
 
-	implementation(size_t output_channels, size_t input_channels, size_t output_sample_rate, size_t input_sample_rate, AVSampleFormat output_sample_format, AVSampleFormat input_sample_format)
+	implementation(int output_channels, int input_channels, int output_sample_rate, int input_sample_rate, AVSampleFormat output_sample_format, AVSampleFormat input_sample_format)
 		: output_channels_(output_channels)
 		, output_sample_format_(output_sample_format)
 		, input_channels_(input_channels)
@@ -89,7 +89,7 @@ struct audio_resampler::implementation
 			auto ret = audio_resample(resampler_.get(),
 									  reinterpret_cast<short*>(buffer2_.data()), 
 									  reinterpret_cast<short*>(data.data()), 
-									  data.size() / (av_get_bytes_per_sample(input_sample_format_) * input_channels_)); 
+									  static_cast<int>(data.size()) / (av_get_bytes_per_sample(input_sample_format_) * input_channels_)); 
 			buffer2_.resize(ret * av_get_bytes_per_sample(output_sample_format_) * output_channels_);
 			std::swap(data, buffer2_);
 		}
@@ -99,7 +99,7 @@ struct audio_resampler::implementation
 };
 
 
-audio_resampler::audio_resampler(size_t output_channels, size_t input_channels, size_t output_sample_rate, size_t input_sample_rate, AVSampleFormat output_sample_format, AVSampleFormat input_sample_format)
+audio_resampler::audio_resampler(int output_channels, int input_channels, int output_sample_rate, int input_sample_rate, AVSampleFormat output_sample_format, AVSampleFormat input_sample_format)
 				: impl_(new implementation(output_channels, input_channels, output_sample_rate, input_sample_rate, output_sample_format, input_sample_format)){}
 std::vector<int8_t, tbb::cache_aligned_allocator<int8_t>> audio_resampler::resample(std::vector<int8_t, tbb::cache_aligned_allocator<int8_t>>&& data){return impl_->resample(std::move(data));}
